% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportLoom.R
\docType{methods}
\name{export,LoomExperiment,loomFile,ANY-method}
\alias{export,LoomExperiment,loomFile,ANY-method}
\title{Method for exporting to .loom files}
\usage{
\S4method{export}{LoomExperiment,loomFile,ANY}(object, con,
  matrix = assayNames(object)[1], rownames_attr = "rownames",
  colnames_attr = "colnames")
}
\arguments{
\item{object}{LoomExperiment the object that is to be exported.}

\item{matrix}{character(1) indicating the name of the matrix in the .loom
file.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the LoomExperiment
object's rownames.}

\item{file}{character(1) indicating the file path to the
that is to be imported.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the LoomExperiment
object's rownames.}
}
\value{
LoomExperiment contained the information from the .loom file.
}
\description{
A function for exporting a \code{LoomExperiment} object to a \code{.loom}
 con.
}
\examples{
temp_con <- tempcon(conext=".h5")
.exportLoom(le, temp_con, rownames_attr="id", colnames_attr="id")
rhdf5::h5ls(temp_con)
rhdf5::h5dump(temp_con)
}
