% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLoomExperimentFromLoom.R
\docType{methods}
\name{export_loom,LoomExperiment-method}
\alias{export_loom,LoomExperiment-method}
\title{Method for exporting to .loom files}
\usage{
\S4method{export_loom}{LoomExperiment}(object, file,
  matrix = assayNames(object)[1], rownames_attr = "rownames",
  colnames_attr = "colnames")
}
\arguments{
\item{object}{LoomExperiment the object that is to be exported.}

\item{file}{character(1) indicating the file path to the
that is to be imported.}

\item{matrix}{character(1) indicating the name of the matrix in the .loom
file.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the LoomExperiment
object's rownames.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the LoomExperiment
object's rownames.}
}
\value{
LoomExperiment contained the information from the .loom file.
}
\description{
A function for exporting a \code{LoomExperiment} object to a \code{.loom}
 file.
}
\examples{
temp_file <- tempfile(fileext=".h5")
export_loom(le, temp_file, rownames_attr="id", colnames_attr="id")
rhdf5::h5ls(temp_file)
rhdf5::h5dump(temp_file)
}
