% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importLoom.R
\docType{methods}
\name{import.loom,ANY-method}
\alias{import.loom,ANY-method}
\title{Function for importing .loom files}
\usage{
\S4method{import.loom}{ANY}(con, ..., rownames_attr = NULL,
  colnames_attr = NULL)
}
\arguments{
\item{con}{character(1) a character vector indicating the file path to the
that is to be imported.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the LoomExperiment
object's rownames.}

\item{rownames_attr}{character indicating the row
attributes in the .loom con that are to be designated as the LoomExperiment
object's rownames.}
}
\value{
LoomExperiment contained the information from the .loom file.
}
\description{
a function for importing a \code{.loom} con as a \code{loomexperiment}.
}
\examples{
test_con <- system.con(
     package="loomexperiment", "extdata", "example.loom"
)
le <- importLoom(test_con)
le
}
