% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLoomExperimentFromLoom.R
\name{import_loom}
\alias{import_loom}
\title{Function for importing .loom files}
\usage{
import_loom(file, rownames_attr = NULL, colnames_attr = NULL)
}
\arguments{
\item{file}{character(1) a character vector indicating the file path to the
that is to be imported.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the loomexperiment
object's rownames.}

\item{rownames_attr}{character indicating the row
attributes in the .loom file that are to be designated as the loomexperiment
object's rownames.}
}
\value{
loomexperiment contained the information from the .loom file.
}
\description{
a function for importing a \code{.loom} file as a \code{loomexperiment}.
}
\examples{
test_file <- system.file(
     package="loomexperiment", "extdata", "example.loom"
)
le <- import_loom(test_file)
le
}
