% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedLoomExperiment-class.R
\docType{class}
\name{RangedLoomExperiment-class}
\alias{RangedLoomExperiment-class}
\title{RangedLoomExperiment}
\description{
A class that helps facilitate the transition of SummarizedExperiment objects
to .loom files and vise versa.
}
\section{Slots}{

\describe{
\item{\code{colGraphs}}{A LoomGraphs object containing the colGraphs information}

\item{\code{rowGraphs}}{A LoomGraphs containing the rowGraphs information}
}}

\author{
Daniel Van Twisk
}
