% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoomGraph-class.R
\docType{methods}
\name{[<-,LoomGraph,ANY,missing,numeric-method}
\alias{[<-,LoomGraph,ANY,missing,numeric-method}
\title{importFrom plyr mapvalues}
\usage{
\S4method{[}{LoomGraph,ANY,missing,numeric}(x, i, j, ...) <- value
}
\description{
importFrom plyr mapvalues
}
