% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedLoomExperiment-class.R
\docType{class}
\name{SummarizedLoomExperiment-class}
\alias{SummarizedLoomExperiment-class}
\title{SummarizedLoomExperiment}
\description{
A class that helps facilitate the transition of SummarizedExperiment objects
to .loom files and vise versa.
}
\section{Slots}{

\describe{
\item{\code{colGraphs}}{A SimpleList containing the colGraphs information}

\item{\code{rowGraphs}}{A SimpleList containing the rowGraphs information}
}}

\author{
Daniel Van Twisk
}
