% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoomExperiment-class.R
\docType{methods}
\name{LoomExperiment}
\alias{LoomExperiment}
\alias{colGraphs,LoomExperiment-method}
\alias{colGraphs<-,LoomExperiment-method}
\alias{rowGraphs,LoomExperiment-method}
\alias{rowGraphs<-,LoomExperiment-method}
\alias{[,LoomExperiment,ANY,ANY,ANY-method}
\title{The LoomExperiment representation class}
\usage{
LoomExperiment(..., colGraphs = LoomGraphs(), rowGraphs = LoomGraphs())

\S4method{colGraphs}{LoomExperiment}(x, ...)

\S4method{colGraphs}{LoomExperiment}(x, ...) <- value

\S4method{rowGraphs}{LoomExperiment}(x, ...)

\S4method{rowGraphs}{LoomExperiment}(x, ...) <- value

\S4method{[}{LoomExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
SummarizedExperiment-like class that facilitate the transition of 
SummarizedExperiment, RangedSummarizedExperiment, and
SingleCellExperiment objects to storage or retrieval in the loom format.
}
\section{Slots}{

\describe{
\item{\code{int_elementMetaData}}{DataFrame. Mirrors int_elementMetadata from
SingleCellExperiment.}

\item{\code{int_colData}}{DataFrame. Mirrors int_colData from SingleCellExperiment.}

\item{\code{int_metadata}}{List. Mirrors int_metadata from SingleCellExperiment.}

\item{\code{reducedDims}}{SimpleList of matrices. Mirrors reducedDims from
SingleCellExperiment.}

\item{\code{colGraphs}}{LoomGraphs containing the loom file's colGraphs
information. Optional.}

\item{\code{rowGraphs}}{LoomGraphs containing the loom file's rowGraphs
information. Optional.}
}}

\examples{

 library(SingleCellExperiment)
 counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
 sce <- SingleCellExperiment(assays = list(counts = counts))
 le <- LoomExperiment(sce)
 le

 ## colGraphs of LoomExperiment object
 colGraphs(le)

 ## replace colGraphs of LoomExperiment object
 lg <- LoomGraph(a=c(1, 2, 3), b=(3, 2, 1))
 colGraphs(le) <- lg

 ## rowGraphs of LoomExperiment object
 rowGraphs(le)

 ## replace rowGraphs of LoomExperiment object
 lg <- LoomGraph(a=c(1, 2, 3), b=(3, 2, 1))
 rowGraphs(le) <- lg

}
\author{
Daniel Van Twisk
}
