% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoomExperiment-class.R
\docType{class}
\name{LoomExperiment-class}
\alias{LoomExperiment-class}
\title{LoomExperiment}
\description{
A class that helps facilitate the transition of SummarizedExperiment objects
to .loom files and vise versa.
}
\section{Slots}{

\describe{
\item{\code{colGraph}}{A SimpleList containing the colGraph information}

\item{\code{rowGraph}}{A SimpleList containing the rowGraph information}
}}

\author{
Daniel Van Twisk
}
