% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoomGraph-class.R
\docType{methods}
\name{LoomGraph}
\alias{LoomGraph}
\alias{[,LoomGraph,ANY,missing,ANY-method}
\title{LoomGraph class for LoomExperiment}
\usage{
LoomGraph(...)

\S4method{[}{LoomGraph,ANY,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
The LoomGraph class is a DataFrame-like class that represents a
 weighted or unweighted graph from the loom file format.
}
\details{
The LoomGraph object must have two or three columns. If the graph
 has two columns, it is an unweighted graph and must have columns named 'a'
 and 'b'. If the graph has three columns, it is treated as a weighted graph
 and must have column names 'a', 'b', and 'w'.  Each row in the LoomGraph
 represents an edge between two column/row numbers and may have a weight.
}
\examples{
 ## create new LoomGraph
 lg <- LoomGraph(a=c(1, 2, 3), b=c(3, 2, 1), w=c(5, 4, 1))
 lg

}
