% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoomGraph-class.R
\docType{methods}
\name{LoomGraphs}
\alias{LoomGraphs}
\alias{[<-,LoomGraph,ANY,missing,numeric-method}
\title{LoomGraphs class for LoomExperiment}
\usage{
LoomGraphs(...)

\S4method{[}{LoomGraph,ANY,missing,numeric}(x, i, j, ...) <- value
}
\description{
The LoomGraphs class contain multiple LoomGraph objects to
 display either colGraph or rowGraph data from the loom file format.

importFrom plyr mapvalues
}
\examples{
 ## create new LoomGraphs
 lg1 <- LoomGraph(a=c(1, 2, 3), b=c(3, 2, 1), w=c(3, 4, 5))
 lg2 <- LoomGraph(a=c(3, 5, 2), b=c(5, 3, 3))
 lgs <- LoomGraphs(lg1, lg2)
 lgs

}
